requirements.txt
faiss_cpu==1.9.0.post1
mistralai==1.3.1
numpy==2.2.2
Requests==2.32.3
tqdm==4.67.1
